Outsource GET request with session cookie
