+++ src/test/kotlin/de/livepoll/api/cucumber/stepdefinitions/PollStepDefinitions.kt
@@ -10,26 +10,14 @@ import io.cucumber.java.en.Given
 import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
 import org.assertj.core.api.Assertions.assertThat
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.*
 import org.springframework.web.client.exchange
 import java.sql.Date
 
-class PollStepDefinitions(userRepository: UserRepository) : CucumberIntegrationTestContext(userRepository) {
-    @Autowired
-    private lateinit var pollRepository: PollRepository
-
+class PollStepDefinitions(userRepository: UserRepository, val pollRepository: PollRepository) : CucumberIntegrationTestContext(userRepository) {
     private val POLLS_ENDPOINT = "/v0/users/${testUser.id}/polls"
     private val CREATE_POLL_ENDPOINT = "/v0/users/${testUser.id}/poll"
-    var sessionCookie: String = ""
     lateinit var myPolls: ArrayList<PollDtoOut>
-
-    @Given("I am logged in as test user")
-    fun beLoggedInAsSpecificUser() {
-        val (status, sessionCookie) = logInWithTestUser()
-        this.sessionCookie = sessionCookie
-        assertThat(status).isEqualTo(HttpStatus.OK)
-    }
 
     @Given("I have no polls created yet")
     fun deleteAllPolls() {
@@ -52,7 +40,7 @@ class PollStepDefinitions(userRepository: UserRepository) : CucumberIntegrationT
         // request body params & headers
         val pollPostRequest = PollDtoIn(pollName, Date(0), Date(0))
         val headers = HttpHeaders()
-        headers["Cookie"] = sessionCookie
+        headers["Cookie"] = SessionCookieUtil.sessionCookie
         val requestEntity: HttpEntity<PollDtoIn> = HttpEntity(pollPostRequest, headers)
 
         // make request
@@ -73,18 +61,7 @@ class PollStepDefinitions(userRepository: UserRepository) : CucumberIntegrationT
     @And("I retrieve my polls")
     fun retrieveMyPolls() {
         val url = "${SERVER_URL}:$port$POLLS_ENDPOINT"
-        // request body params & headers
-        val headers = HttpHeaders()
-        headers["Cookie"] = sessionCookie
-        val requestEntity: HttpEntity<String> = HttpEntity("", headers)
-
-        // make request
-        // val pollResponseEntity: ResponseEntity<Array<Poll>> = restTemplate.getForEntity(url)
-        val pollResponseEntity: ResponseEntity<Array<PollDtoOut>> = restTemplate.exchange(
-                url,
-                HttpMethod.GET,
-                requestEntity
-        )
+        val pollResponseEntity = makeGetRequestWithSessionCookie<Array<PollDtoOut>>(url, SessionCookieUtil.sessionCookie)
         assertThat(pollResponseEntity.statusCode).isEqualTo(HttpStatus.OK)
         assertThat(pollResponseEntity.body).isNotNull
         for (poll in pollResponseEntity.body!!) {
